更新 main.py：整合計費、地圖圖片、Google Maps 連結

+++ main.py
@@ -41,6 +41,7 @@ def get_route(data: RouteRequest):
         if not directions:
             raise HTTPException(status_code=404, detail="無法取得路線")
 
+        # 所有路線的距離與時間
         routes = []
         best_duration = float("inf")
         best_leg = None
@@ -56,11 +57,13 @@ def get_route(data: RouteRequest):
                 best_duration = duration_min
                 best_leg = leg
 
+        # 最佳路線計費
         best_distance = best_leg["distance"]["value"] / 1000
-        fee = round(best_distance * 2 * 3)
+        fee = int(round(best_distance * 2 * 3))  # 四捨五入後取整數
         today = datetime.today().strftime("%Y-%m-%d")
         report = f"{today} {data.origin}-{data.destination}【自行開車 {best_distance:.1f}(公里數)*3(元/公里)*2(來回)={fee}(費用)】"
 
+        # 靜態地圖圖片
         static_map_url = (
             "https://maps.googleapis.com/maps/api/staticmap?" +
             urlencode({
@@ -86,7 +89,7 @@ def get_route(data: RouteRequest):
             "fee": fee,
             "report": report,
             "map_url": map_url,
-            "map_link": map_link  # ✅ 已加入 Google Maps 連結
+            "map_link": map_link
         }
 
     except Exception as e:
